+++ src/SeleniumLibrary/keywords/element.py
@@ -384,6 +384,26 @@ class ElementKeywords(LibraryComponent):
         return self.find_element(locator).get_attribute(attribute)
 
     @keyword
+    def element_attribute_value_should_be(self, locator, attribute, expected, message=None):
+        """Verifies element identified by ``locator`` contains expected attribute value.
+
+        See the `Locating elements` section for details about the locator
+        syntax.
+
+        Example:
+        `Element Attribute Value Should Be` | css:img | href | value
+
+        New in SeleniumLibrary 3.2.
+        """
+        current_expected = self.find_element(locator).get_attribute(attribute)
+        if current_expected != expected:
+            if is_noney(message):
+                message = ("Element '%s' attribute should have value '%s' but "
+                          "its value was '%s'." % (locator, expected, current_expected))
+            raise AssertionError(message)
+        self.info("Element '%s' attribute '%s' contains value '%s'." % (locator, attribute, expected))
+
+    @keyword
     def get_horizontal_position(self, locator):
         """Returns horizontal position of element identified by ``locator``.